refactor: move SpinalSimUsbHostTester
